move solid angle tests back into normalization tests and remove undefined provider in beam_center_finder
